% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as_treedata.R
\name{as_treedata}
\alias{as_treedata}
\alias{as_treedata,eclust-method}
\title{transform the eclust class to treedata class}
\usage{
as_treedata(.data, bycol = TRUE, assay = 1, longer = FALSE, metadata = FALSE)

\S4method{as_treedata}{eclust}(.data, bycol = TRUE, assay = 1, longer = FALSE, metadata = FALSE)
}
\arguments{
\item{.data}{R object, here is eclust class}

\item{bycol}{cluster by column}

\item{assay}{assay of the eclust that is to be saved in treedata, default is 1}

\item{longer}{transform the assay to longer type}

\item{metadata}{add the metadata of the eclust}
}
\value{
treedata class
}
\description{
transform the eclust class to treedata class
}
\examples{
data(ec)
tree <- as_treedata(ec,longer = TRUE)
}
