% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\name{eclust-accessors}
\alias{eclust-accessors}
\alias{colhclust}
\alias{colhclust,eclust-method}
\alias{colhclust,eclust}
\alias{rowhclust}
\alias{rowhclust,eclust-method}
\alias{rowhclust,eclust}
\alias{colhclust<-}
\alias{colhclust<-,eclust,hclust_or_NULL-method}
\alias{colhclust<-,eclust,hclust_or_NULL}
\alias{rowhclust<-}
\alias{rowhclust<-,eclust,hclust_or_NULL-method}
\alias{rowhclust<-,eclust,hclust_or_NULL}
\title{eclust accessors}
\usage{
colhclust(x, ...)

\S4method{colhclust}{eclust}(x, ...)

rowhclust(x, ...)

\S4method{rowhclust}{eclust}(x, ...)

colhclust(x, ...) <- value

\S4method{colhclust}{eclust,hclust_or_NULL}(x, ...) <- value

rowhclust(x, ...) <- value

\S4method{rowhclust}{eclust,hclust_or_NULL}(x, ...) <- value
}
\arguments{
\item{x}{R object, here is eclust class}

\item{...}{additional parameters.}

\item{value}{hclust or NULL}
}
\description{
eclust accessors
}
