% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{eclust-class}
\alias{eclust-class}
\title{eclust class}
\description{
eclust class
}
\section{Slots}{

\describe{
\item{\code{rowhclust}}{hclust object to store the result of hierarchical cluster based on
the row index of SingleCellExperiment.}

\item{\code{colhclust}}{hclust object to store the result of hierarchical cluster based on
the column index of SingleCellExperiment.}
}}

