% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eclust.R
\name{eclust}
\alias{eclust}
\alias{eclust,SummarizedExperiment-method}
\alias{eclust,SummarizedExperiment}
\alias{eclust,RangedSummarizedExperiment-method}
\alias{eclust,RangedSummarizedExperiment}
\title{Hierarchical cluster analysis for the sample or features with SummarizedExperiment class}
\usage{
eclust(
  .data,
  .assay,
  dist.method = "euclidean",
  hclust.method = "average",
  direction = "col",
  ...
)

\S4method{eclust}{SummarizedExperiment}(
  .data,
  .assay,
  dist.method = "euclidean",
  hclust.method = "average",
  direction = "col",
  ...
)

\S4method{eclust}{RangedSummarizedExperiment}(
  .data,
  .assay,
  dist.method = "euclidean",
  hclust.method = "average",
  direction = "col",
  ...
)
}
\arguments{
\item{.data}{R object, here is SummarizedExperiment class}

\item{.assay}{the name or index of the assays in SummarizedExperiment class,
default is 1.}

\item{dist.method}{the method to calculate the distance, default is 'euclidean'.}

\item{hclust.method}{the agglomeration method for the hierarchical cluster,
default is 'average'.}

\item{direction}{sample ('col') or features ('row') to preform the hierarchical
cluster analysis, default is 'col'.}

\item{...}{additional parameters}
}
\value{
eclust class
}
\description{
Hierarchical cluster analysis for the sample or features with SummarizedExperiment class
}
